Fix SQL query tooltips showing up when dorpdown menu is open or  modal is open for non active query
